Renamed Http Gem to Http.rb
