+++ lib/webmock/http_lib_adapters/http_rb/streamer.rb
